Redactions and corrections
after reading the proj again, the server fifo is given by the user when the server and any client is created, so we cant have it defined.

Fixed parts of the pub and removed the send_msg from util_funcs. Still confused as to whether we have to truncate normal messages from the pub, with what and when.